Disable write batching for PSQL
Problem: batches of transactions are accumulated before an EOF of trace
file is encountered. Given files are written in parallel by the fetcher,
this may lead to potentially indefinite accumulation of batches without
commitment.

In practice this leads to accumulation of large batches and then
contention to submit these batches to the SQL database (as 200
concurrent writers attempt to make a commitment).

Solution: disable batching for Postgresql (as it was originally
introduce as a necessary optimization for sqlite).